pyproject.toml
[tool.poetry]
name = "test_first"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
django = ">2.2,<3.0"
djangorestframework = "^3.12.4"
django-import-export = "^2.5.0"
psycopg2-binary = "^2.8.6"
gunicorn = "^20.1.0"
django-environ = "^0.4.5"
django-redis = "^4.12.1"
django-money = "^1.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
